// src/app/models/requisicao.model.ts
export interface Requisicao {
  codigoEmpresa: string;
  codigoProduto: string;
  codigoDerivacao: string;
  codigoDepartamento: string;
  dataMovimentada: string;
  sequenciaMovimento: string;
  codigoTransferencia: string;
  numeroDocumento: string;
  observacaoMovimento: string;
  quatidadeMovimento: string;
  valorMovimentado: string;
  usuarioResponsavel: string;
  usuarioRecebedor: string;
}